// src/api/generated/model/cityUpdate.ts
/**
 * Generated by orval v6.27.1 🍺
 * Do not edit manually.
 * API Geo
 * OpenAPI spec version: 0.1.0
 */
import type { CityUpdateCountryId } from "./cityUpdateCountryId";
import type { CityUpdateDeletedAt } from "./cityUpdateDeletedAt";
import type { CityUpdateIata } from "./cityUpdateIata";
import type { CityUpdateLatitude } from "./cityUpdateLatitude";
import type { CityUpdateLongitude } from "./cityUpdateLongitude";
import type { CityUpdateName } from "./cityUpdateName";
import type { CityUpdateNeedAutomaticUpdate } from "./cityUpdateNeedAutomaticUpdate";
import type { CityUpdateOsmId } from "./cityUpdateOsmId";
import type { CityUpdateOsmType } from "./cityUpdateOsmType";
import type { CityUpdateRegionId } from "./cityUpdateRegionId";
import type { CityUpdateTimezone } from "./cityUpdateTimezone";

export interface CityUpdate {
  country_id?: CityUpdateCountryId;
  deleted_at?: CityUpdateDeletedAt;
  iata?: CityUpdateIata;
  latitude?: CityUpdateLatitude;
  longitude?: CityUpdateLongitude;
  name?: CityUpdateName;
  need_automatic_update?: CityUpdateNeedAutomaticUpdate;
  osm_id?: CityUpdateOsmId;
  osm_type?: CityUpdateOsmType;
  region_id?: CityUpdateRegionId;
  timezone?: CityUpdateTimezone;
}